# docs/swagger.yaml
basePath: /
definitions:
  dto_auth.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      role:
        example: member
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  dto_auth.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  dto_auth.ProfileData:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 1
        type: integer
      role:
        example: member
        type: string
    type: object
  dto_auth.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    required:
    - refresh_token
    type: object
  dto_auth.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  utils.DataResponseDTO:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  utils.ErrorResponseDTO:
    properties:
      message:
        example: An error occurred
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.MessageResponseDTO:
    properties:
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: A production-ready RESTful backend with JWT authentication (Access
    + Refresh tokens), Role-Based Access Control, and Clean Architecture.
  title: Backend Template API
  version: 1.1.0
paths:
  /api/v1/auth/health:
    get:
      description: Check if the auth service is running and view role hierarchy.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.MessageResponseDTO'
      summary: Health check
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password. Returns access token
        (6h) and refresh token (24h).
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.DataResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto_auth.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the refresh token to invalidate the session.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.MessageResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /api/v1/auth/profile:
    get:
      description: Retrieve the authenticated user's profile information.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.DataResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto_auth.ProfileData'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a valid refresh token. Implements
        token rotation (old refresh token is revoked).
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.DataResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto_auth.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password. Default role
        is "member".
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.DataResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto_auth.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
      summary: Register a new user
      tags:
      - Auth
  /api/v1/auth/users:
    get:
      description: Retrieve a list of all users. Requires admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.MessageResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"